+++ tests/test_predict.py
@@ -1,5 +1,6 @@
 import gym
 import pytest
+import torch as th
 
 from stable_baselines3 import A2C, DQN, PPO, SAC, TD3
 from stable_baselines3.common.vec_env import DummyVecEnv
@@ -30,7 +31,11 @@ def test_auto_wrap(model_class):
 
 @pytest.mark.parametrize("model_class", MODEL_LIST)
 @pytest.mark.parametrize("env_id", ["Pendulum-v0", "CartPole-v1"])
-def test_predict(model_class, env_id):
+@pytest.mark.parametrize("device", ["cpu", "cuda", "auto"])
+def test_predict(model_class, env_id, device):
+    if device == "cuda" and not th.cuda.is_available():
+        pytest.skip("CUDA not available")
+
     if env_id == "CartPole-v1":
         if model_class in [SAC, TD3]:
             return
@@ -38,7 +43,7 @@ def test_predict(model_class, env_id):
         return
 
     # test detection of different shapes by the predict method
-    model = model_class("MlpPolicy", env_id)
+    model = model_class("MlpPolicy", env_id, device=device)
     env = gym.make(env_id)
     vec_env = DummyVecEnv([lambda: gym.make(env_id), lambda: gym.make(env_id)])